pyproject.toml
# pyproject.toml

[build-system]
  requires = ["setuptools>=65.0.0"]
  build-backend = "setuptools.build_meta"

[project]
  name = "theodorblom.com"
  version = "1.0.0"
  authors = [
    { name = "Theodor Blom", email = "me@theodorblom.com" },
  ]
  description = "A personal website deployed as a Flask app instance on a VPS."
  readme = "README.md"
  requires-python = ">=3.8,<3.9"
  classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Framework :: Flask",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: Unix",
  ]
  dependencies = [
    "Flask >= 3.0.1, < 4",
    "Flask-login >= 0.6.3, < 1",
    "Flask-sqlalchemy >= 3.1.1, < 4",
    "Flask-Migrate >= 4.0.5, < 5",
    "Flask-Sitemap >= 0.4.0, < 1",
    "python-slugify >= 8.0.2, < 9",
    "markdown >= 3.5.2, < 4",
  ]

[project.optional-dependencies]
  dev = [
    "pytest >= 7.4.4, < 8",
    "pytest-cov >= 4.1.0, < 5",
  ]

[tool.setuptools.package-data]
  "website.templates" = ["**/**"]
  "website.static" = ["**/**"]

[tool.pytest.ini_options]
  addopts = "-v --cov=./src"
  testpaths = "tests"

[project.urls]
  homepage = "https://github.com/Theeoi/theodorblom.com"
